refactor(nr_ext): move PREDICT handler to nr_ext
